Imported the full folder structure from osTicket.
This makes it easier for people to get an installation up and running quicker if they are cloning this branch.

This commit does NOT include the database changes in previous commit of this branch. I will include them in the next commit to make it easier for people to install this branch.